Samples modified for demonstration of FromSVG() function (#65)
* Add SVGConvertOptions

Added SVGConvertOptions to swig include files for Python, PHP, Go and
Ruby. This is needed for the new Convert::FromSVG() function.

* PDFLIB-177: PHP wrapper test updated also to run Convert::FromSVG()

* PDFLIB-178: Python wrapper test updated also to run Convert::FromSVG()

* PDFLIN-176: GO wrapper test updated also to run Convert::FromSVG()

* Fix SVGConvertOptions.h includes

Moved the inclusion of the SVGConvertOptions.h to the correct place
in the inclusion sequence before Convert.h. This fixes the error that
FromSVG() with an SVGConvertOptions argument caused an error in
all wrappers because no matching overload was found.

* Reverting samples modifications by Igor

Reverting the modifications to the samples by Igor, as these must
not yet become effective until the integrated PDFlib build
is enabled in general.

* PDFLIB-176: GO wrapper test updated also to run Convert::FromSVG()

* PDFLIB-178: Python wrapper test updated also to run Convert::FromSVG()

* PDFLIB-177: PHP wrapper test updated also to run Convert::FromSVG()

* PDFLIB-178: Ruby sample for SVG import

* Use FromSVG() explicitly in Ruby ConvertTest

* Fix wrong method name

* PDFLIB-176: GO SVG converter test updated

* PDFLIB-177: PHP SVG converter test updated

* PDFLIB-178: Python SVG converter test updated

* Fix creation of SVGConvertOptions object for Go

* Fix errors in PHP ConvertTest sample

* Remove SVGOptions from Python ConvertTest

According to David a typical usage by customers should be
demonstrated.

* Omit SVGOptions from ConvertTest Ruby sample

* Omit SVGOptions from PHP ConvertTest sample

* Omit SVGOptions from Go ConvertTest sample

---------

Co-authored-by: Igor Repinetski <irepinetski@pdftron.com>